backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
pillow==9.5.0
tqdm==4.65.0
scikit-learn==1.3.0
python-multipart==0.0.6
aiofiles==23.1.0
numpy==1.26.0
opencv-python==4.9.0.80
supabase==2.6.0